feat: refactor navigation
